+++ python/cudf/cudf/pandas/scripts/conftest-patch.py
@@ -3419,7 +3419,6 @@ NODEIDS_THAT_FAIL_WITH_CUDF_PANDAS = {
     "tests/extension/test_arrow.py::TestArrowArray::test_setitem_frame_2d_values[uint32]",
     "tests/extension/test_arrow.py::TestArrowArray::test_setitem_frame_2d_values[uint64]",
     "tests/extension/test_arrow.py::TestArrowArray::test_setitem_frame_2d_values[uint8]",
-    "tests/extension/test_arrow.py::TestArrowArray::test_setitem_with_expansion_row[decimal128(7, 3)]",
     "tests/extension/test_arrow.py::TestArrowArray::test_to_numpy[decimal128(7, 3)]",
     "tests/extension/test_arrow.py::TestArrowArray::test_to_numpy[double]",
     "tests/extension/test_arrow.py::TestArrowArray::test_to_numpy[duration[ms]]",
@@ -10629,7 +10628,6 @@ NODEIDS_THAT_FAIL_WITH_CUDF_PANDAS = {
     "tests/indexing/test_partial.py::TestPartialSetting::test_loc_with_list_of_strings_representing_datetimes[DataFrame-idx1-labels1-expected_idx1]",
     "tests/indexing/test_partial.py::TestPartialSetting::test_loc_with_list_of_strings_representing_datetimes[Series-idx1-labels1-expected_idx1]",
     "tests/indexing/test_partial.py::TestPartialSetting::test_partial_setting_frame",
-    "tests/indexing/test_partial.py::TestPartialSetting::test_partial_setting_mixed_dtype",
     "tests/indexing/test_scalar.py::TestMultiIndexScalar::test_multiindex_at_set",
     "tests/indexing/test_scalar.py::TestScalar::test_at_set_ints_other[floats-series]",
     "tests/indexing/test_scalar.py::TestScalar::test_at_set_ints_other[ints-series]",
